(Important) Update preprocess.R
The whole logic of preprocessing was updated. When plotting boxplots for each sample, some samples clearly have much shorter boxes, with tails truncated. This phenomenon was probably due to the not so robust imputation method (imputing with the Ratio of Detection). This made me consider using a more powerful imputation method, so I turned to KNN. However, considering the mechanism of KNN imputation, it makes more sense to impute on normalized data. So the data must be normalized with missing values. I used median quotient normalization before, which depends on the median for each sample. I don't know about the details of how glycanr package deals with missing values, but I assume that the median of a sample will be affected by missing values, unless the sample has missing values for less than 50% glycans and all missing values were firstly imputed with some value below the median. Based on these thinkings, the new preprocessing workflow is as follows:

1. Filtering glycans: remove glycans with missing values in more than 50% sample.
2. Filtering samples: remove samples with missing values for more than 50% glycans.
3. Normalization: first impute with min / 2 for each sample, then normalize by median quotient, finally put those missing values back.
4. Imputation: Impute with KNN.